+++ tests/ByteSync.Client.IntegrationTests/TestHelpers/Http/FlakyOnceHandler.cs
@@ -1,5 +1,4 @@
 using System.Net;
-using System.Net.Http;
 
 namespace ByteSync.Client.IntegrationTests.TestHelpers.Http;
