+++ tests/ByteSync.Client.IntegrationTests/TestHelpers/Http/IFlakyCounter.cs
@@ -1,0 +1,6 @@
+﻿namespace ByteSync.Client.IntegrationTests.TestHelpers.Http;
+
+public interface IFlakyCounter
+{
+    bool ShouldFail(HttpMethod method);
+}
